move option tally logic to PollNoteViewModel
